r/aws_redshift_snapshot_schedule_association: Move functions around.
