+++ internal/service/redshift/status.go
@@ -76,22 +76,6 @@ func statusClusterAqua(ctx context.Context, conn *redshift.Redshift, id string) 
 	}
 }
 
-func statusScheduleAssociation(ctx context.Context, conn *redshift.Redshift, id string) retry.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		_, output, err := FindScheduleAssociationById(ctx, conn, id)
-
-		if tfresource.NotFound(err) {
-			return nil, "", nil
-		}
-
-		if err != nil {
-			return nil, "", err
-		}
-
-		return output, aws.StringValue(output.ScheduleAssociationState), nil
-	}
-}
-
 func statusEndpointAccess(ctx context.Context, conn *redshift.Redshift, name string) retry.StateRefreshFunc {
 	return func() (interface{}, string, error) {
 		output, err := FindEndpointAccessByName(ctx, conn, name)
